Move atlas/fortran to atlas_f

+++ src/atlas/actions/GenerateMesh.cc
@@ -24,8 +24,8 @@ namespace actions {
 Mesh* generate_mesh (const ReducedGrid& rgg)
 {
   ReducedGridMeshGenerator generate;
-  generate.options.set( "nb_parts", (int) eckit::mpi::size() );
-  generate.options.set( "part"    , (int) eckit::mpi::rank() );
+  generate.options.set( "nb_parts", eckit::mpi::size() );
+  generate.options.set( "part"    , eckit::mpi::rank() );
   return generate(rgg);
 }
